@@ -15,7 +15,8 @@
   "dependencies": {
     "babel-polyfill": "^6.26.0",
     "express": "^4.16.3",
-    "react": "^16.5.2",
+    "react": "^16.10.2",
+    "ramda": "^0.26.1",
     "react-dom": "^16.5.2"
   },
   "devDependencies": {
@@ -30,12 +31,15 @@
     "css-loader": "^2.0.0",
     "eslint": "^5.0.0",
     "eslint-config-airbnb": "^17.0.0",
+    "eslint-config-prettier": "^6.4.0",
     "eslint-plugin-import": "^2.11.0",
     "eslint-plugin-jsx-a11y": "^6.0.3",
+    "eslint-plugin-prettier": "^3.1.1",
     "eslint-plugin-react": "^7.7.0",
     "file-loader": "^3.0.0",
     "html-webpack-plugin": "^3.2.0",
     "nodemon": "^1.17.3",
+    "prettier-eslint": "^9.0.0",
     "style-loader": "^0.23.0",
     "url-loader": "^1.0.1",
     "webpack": "^4.5.0",
