@@ -1,23 +1,33 @@
-import React, { Component } from 'react';
-import './app.css';
-import ReactImage from './react.png';
+import React, { useState, useEffect } from 'react'
+import './app.css'
+// import R from 'ramda'
+// import Effects from './Effects'
 
-export default class App extends Component {
-  state = { username: null };
+export default function Example() {
+  // const [username, setUsername] = useState(null)
+  // const { getUsername } = Effects
 
-  componentDidMount() {
-    fetch('/api/getUsername')
-      .then(res => res.json())
-      .then(user => this.setState({ username: user.username }));
-  }
+  // useEffect(() => getUsername(setUsername))
 
-  render() {
-    const { username } = this.state;
+  // const Message = () => (username ? <h1>{`Hello ${username}`}</h1> : <h1>Loading.. please wait!</h1>)
+
+  const Message = props => {
+    const { name } = props
     return (
-      <div>
-        {username ? <h1>{`Hello ${username}`}</h1> : <h1>Loading.. please wait!</h1>}
-        <img src={ReactImage} alt="react" />
-      </div>
-    );
+      <ul className="message">
+        {name.split('').map(letter => (
+          <li>{letter}</li>
+        ))}
+      </ul>
+    )
   }
+
+  return (
+    <React.Fragment>
+      <Message name="CARSON POWERS" />
+      <video id="vid" playsInline autoPlay loop muted>
+        <source src="../assets/webm/city-sideways.webm" type='video/webm; codecs="vp8, vorbis"' />
+      </video>
+    </React.Fragment>
+  )
 }
