@@ -1,9 +1,9 @@
-const express = require('express');
-const os = require('os');
+const express = require('express')
+const os = require('os')
 
-const app = express();
+const app = express()
 
-app.use(express.static('dist'));
-app.get('/api/getUsername', (req, res) => res.send({ username: os.userInfo().username }));
+app.use(express.static('dist'))
+app.get('/api/getUsername', (req, res) => res.send({ username: os.userInfo().username }))
 
-app.listen(process.env.PORT || 8080, () => console.log(`Listening on port ${process.env.PORT || 8080}!`));
+app.listen(process.env.PORT || 8080, () => console.log(`Listening on port ${process.env.PORT || 8080}!`))
